+++ tests/lib/js/export/functional_test_lib.dart
@@ -1,14 +1,11 @@
-// Copyright (c) 2022, the Dart project authors.  Please see the AUTHORS file
+// Copyright (c) 2025, the Dart project authors.  Please see the AUTHORS file
 // for details. All rights reserved. Use of this source code is governed by a
 // BSD-style license that can be found in the LICENSE file.
 
-// Test various uses of exports that are returned from `createDartExport` or
-// `createJSInteropWrapper`.
-
 import 'dart:js_interop';
+import 'dart:js_interop_unsafe';
 
 import 'package:expect/legacy/minitest.dart'; // ignore: deprecated_member_use_from_same_package
-import 'package:js/js_util.dart';
 
 // Test exporting all vs. only some members.
 // Also test using a non-empty `@JSExport` annotation on a class, which should
@@ -41,18 +38,18 @@ void testExportAll(WrapperCreator creator) {
   var all = creator.createExportAll(dartInstance);
 
   // Verify only the exportable properties exist.
-  expect(hasProperty(all, 'constructor'), false);
-  expect(hasProperty(all, 'factory'), false);
-  expect(hasProperty(all, 'field'), true);
-  expect(hasProperty(all, 'finalField'), true);
-  expect(hasProperty(all, '_getSetField'), true);
-  expect(hasProperty(all, 'getSet'), true);
-  expect(hasProperty(all, 'method'), true);
-  expect(hasProperty(all, 'staticField'), false);
-  expect(hasProperty(all, 'staticMethod'), false);
-  expect(hasProperty(all, 'extensionMethod'), false);
-  expect(hasProperty(all, 'extensionStaticField'), false);
-  expect(hasProperty(all, 'extensionStaticMethod'), false);
+  expect(all.has('constructor'), false);
+  expect(all.has('factory'), false);
+  expect(all.has('field'), true);
+  expect(all.has('finalField'), true);
+  expect(all.has('_getSetField'), true);
+  expect(all.has('getSet'), true);
+  expect(all.has('method'), true);
+  expect(all.has('staticField'), false);
+  expect(all.has('staticMethod'), false);
+  expect(all.has('extensionMethod'), false);
+  expect(all.has('extensionStaticField'), false);
+  expect(all.has('extensionStaticMethod'), false);
 }
 
 class ExportSome {
@@ -82,18 +79,18 @@ void testExportSome(WrapperCreator creator) {
   var some = creator.createExportSome(dartInstance);
 
   // Verify only the properties we marked as exportable exist.
-  expect(hasProperty(some, 'constructor'), false);
-  expect(hasProperty(some, 'factory'), false);
+  expect(some.has('constructor'), false);
+  expect(some.has('factory'), false);
 
-  expect(hasProperty(some, 'field'), true);
-  expect(hasProperty(some, 'finalField'), true);
-  expect(hasProperty(some, 'getSet'), true);
-  expect(hasProperty(some, 'method'), true);
+  expect(some.has('field'), true);
+  expect(some.has('finalField'), true);
+  expect(some.has('getSet'), true);
+  expect(some.has('method'), true);
 
-  expect(hasProperty(some, 'nonExportField'), false);
-  expect(hasProperty(some, 'nonExportFinalField'), false);
-  expect(hasProperty(some, 'nonExportGetSet'), false);
-  expect(hasProperty(some, 'nonExportMethod'), false);
+  expect(some.has('nonExportField'), false);
+  expect(some.has('nonExportFinalField'), false);
+  expect(some.has('nonExportGetSet'), false);
+  expect(some.has('nonExportMethod'), false);
 }
 
 // Test that the properties are forwarded correctly in the object literal.
@@ -101,17 +98,17 @@ void testForwarding(WrapperCreator creator) {
   var dartInstance = ExportAll.constructor();
   var all = creator.createExportAll(dartInstance);
 
-  expect(getProperty(all, 'field'), dartInstance.field);
-  setProperty(all, 'field', 'modified');
+  expect(all['field'], dartInstance.field.toJS);
+  all['field'] = 'modified'.toJS;
   expect(dartInstance.field, 'modified');
 
-  expect(getProperty(all, 'finalField'), dartInstance.finalField);
+  expect(all['finalField'], dartInstance.finalField.toJS);
 
-  expect(getProperty(all, 'getSet'), dartInstance.getSet);
-  setProperty(all, 'getSet', 'modified');
+  expect(all['getSet'], dartInstance.getSet.toJS);
+  all['getSet'] = 'modified'.toJS;
   expect(dartInstance.getSet, 'modified');
 
-  expect(callMethod(all, 'method', []), dartInstance.method());
+  expect(all.callMethod('method'.toJS), dartInstance.method().toJS);
 }
 
 // Test inheritance and overrides for every kind of member.
@@ -136,16 +133,16 @@ void testInheritance(WrapperCreator creator) {
   var dartInheritance = Inheritance();
   var inheritance = creator.createInheritance(dartInheritance);
 
-  expect(getProperty(inheritance, 'field'), dartInheritance.field);
-  setProperty(inheritance, 'field', 'modified');
+  expect(inheritance['field'], dartInheritance.field.toJS);
+  inheritance['field'] = 'modified'.toJS;
   expect(dartInheritance.field, 'modified');
 
-  expect(getProperty(inheritance, 'finalField'), dartInheritance.finalField);
+  expect(inheritance['finalField'], dartInheritance.finalField.toJS);
 
-  expect(getProperty(inheritance, 'getSet'), dartInheritance.getSet);
-  setProperty(inheritance, 'getSet', 'superclassSetter');
+  expect(inheritance['getSet'], dartInheritance.getSet.toJS);
+  inheritance['getSet'] = 'superclassSetter'.toJS;
 
-  expect(callMethod(inheritance, 'method', []), dartInheritance.method());
+  expect(inheritance.callMethod('method'.toJS), dartInheritance.method().toJS);
 }
 
 @JSExport()
@@ -164,19 +161,19 @@ void testOverrides(WrapperCreator creator) {
   var dartOverrides = Overrides();
   var overrides = creator.createOverrides(dartOverrides);
 
-  expect(getProperty(overrides, 'field'), dartOverrides.field);
-  setProperty(overrides, 'field', 'modified');
+  expect(overrides['field'], dartOverrides.field.toJS);
+  overrides['field'] = 'modified'.toJS;
   expect(dartOverrides.field, 'modified');
 
-  expect(getProperty(overrides, 'finalField'), dartOverrides.finalField);
+  expect(overrides['finalField'], dartOverrides.finalField.toJS);
 
-  expect(getProperty(overrides, 'getSet'), dartOverrides.getSet);
-  setProperty(overrides, 'getSet', 'derivedSetter');
+  expect(overrides['getSet'], dartOverrides.getSet.toJS);
+  overrides['getSet'] = 'derivedSetter'.toJS;
 
-  expect(callMethod(overrides, 'method', []), dartOverrides.method());
+  expect(overrides.callMethod('method'.toJS), dartOverrides.method().toJS);
   expect(
-    callMethod(overrides, 'nonOverriddenMethod', []),
-    dartOverrides.nonOverriddenMethod(),
+    overrides.callMethod('nonOverriddenMethod'.toJS),
+    dartOverrides.nonOverriddenMethod().toJS,
   );
 }
 
@@ -205,11 +202,11 @@ void testShadowed(WrapperCreator creator) {
   var dartShadowed = InheritanceShadowed();
   var shadowed = creator.createInheritanceShadowed(dartShadowed);
 
-  expect(hasProperty(shadowed, 'field'), false);
-  expect(hasProperty(shadowed, 'finalField'), false);
-  expect(hasProperty(shadowed, 'getSet'), false);
+  expect(shadowed.has('field'), false);
+  expect(shadowed.has('finalField'), false);
+  expect(shadowed.has('getSet'), false);
 
-  expect(callMethod(shadowed, 'method', []), dartShadowed.method());
+  expect(shadowed.callMethod('method'.toJS), dartShadowed.method().toJS);
 }
 
 // Test that the arity is correct.
@@ -225,30 +222,26 @@ class Arity {
 void testArity(WrapperCreator creator) {
   var arity = creator.createArity(Arity());
 
-  callMethod(arity, 'onePositional', ['']);
+  arity.callMethod('onePositional'.toJS, ''.toJS);
 
-  callMethod(arity, 'twoPositional', ['', '']);
+  arity.callMethod('twoPositional'.toJS, ''.toJS, ''.toJS);
 
-  callMethod(arity, 'oneOptional', []);
-  callMethod(arity, 'oneOptional', ['']);
+  arity.callMethod('oneOptional'.toJS);
+  arity.callMethod('oneOptional'.toJS, ''.toJS);
 
-  callMethod(arity, 'twoOptional', []);
-  callMethod(arity, 'twoOptional', ['']);
-  callMethod(arity, 'twoOptional', ['', '']);
+  arity.callMethod('twoOptional'.toJS);
+  arity.callMethod('twoOptional'.toJS, ''.toJS);
+  arity.callMethod('twoOptional'.toJS, ''.toJS, ''.toJS);
 
-  callMethod(arity, 'onePositionalOneOptional', ['']);
-  callMethod(arity, 'onePositionalOneOptional', ['', '']);
+  arity.callMethod('onePositionalOneOptional'.toJS, ''.toJS);
+  arity.callMethod('onePositionalOneOptional'.toJS, ''.toJS, ''.toJS);
 }
 
 // Test that the transformation occurs in other js_util calls.
 void testNestedJsUtil(WrapperCreator creator) {
-  setProperty(
-    globalThis,
-    'export',
-    creator.createExportAll(ExportAll.constructor()),
-  );
-  expect(hasProperty(globalThis, 'export'), true);
-  expect(hasProperty(getProperty(globalThis, 'export'), 'field'), true);
+  globalContext['export'] = creator.createExportAll(ExportAll.constructor());
+  expect(globalContext.has('export'), true);
+  expect((globalContext['export'] as JSObject).has('field'), true);
 }
 
 void test(WrapperCreator creator) {
@@ -267,39 +260,10 @@ void test(WrapperCreator creator) {
 // parameter to be statically available, we have to use methods that statically
 // declare what class they want wrapped.
 abstract class WrapperCreator {
-  Object createExportAll(ExportAll instance);
-  Object createExportSome(ExportSome instance);
-  Object createInheritance(Inheritance instance);
-  Object createInheritanceShadowed(InheritanceShadowed instance);
-  Object createOverrides(Overrides instance);
-  Object createArity(Arity instance);
-}
-
-class UseCreateDartExport implements WrapperCreator {
-  Object createExportAll(ExportAll instance) => createDartExport(instance);
-  Object createExportSome(ExportSome instance) => createDartExport(instance);
-  Object createInheritance(Inheritance instance) => createDartExport(instance);
-  Object createInheritanceShadowed(InheritanceShadowed instance) =>
-      createDartExport(instance);
-  Object createOverrides(Overrides instance) => createDartExport(instance);
-  Object createArity(Arity instance) => createDartExport(instance);
-}
-
-class UseCreateJSInteropWrapper implements WrapperCreator {
-  JSObject createExportAll(ExportAll instance) =>
-      createJSInteropWrapper(instance);
-  JSObject createExportSome(ExportSome instance) =>
-      createJSInteropWrapper(instance);
-  JSObject createInheritance(Inheritance instance) =>
-      createJSInteropWrapper(instance);
-  JSObject createInheritanceShadowed(InheritanceShadowed instance) =>
-      createJSInteropWrapper(instance);
-  JSObject createOverrides(Overrides instance) =>
-      createJSInteropWrapper(instance);
-  JSObject createArity(Arity instance) => createJSInteropWrapper(instance);
-}
-
-void main() {
-  test(UseCreateDartExport());
-  test(UseCreateJSInteropWrapper());
+  JSObject createExportAll(ExportAll instance);
+  JSObject createExportSome(ExportSome instance);
+  JSObject createInheritance(Inheritance instance);
+  JSObject createInheritanceShadowed(InheritanceShadowed instance);
+  JSObject createOverrides(Overrides instance);
+  JSObject createArity(Arity instance);
 }